+++ backend/Application/Aggregates/Contract/Entities/ModuleReferenceEvent.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using Application.Aggregates.Contract.Types;
 using Application.Api.GraphQL;
 using Application.Api.GraphQL.EfCore;
+using Application.Interop;
 using Concordium.Sdk.Types;
 using HotChocolate;
 using HotChocolate.Types;
@@ -17,6 +18,7 @@ namespace Application.Aggregates.Contract.Entities;
 /// </summary>
 public sealed class ModuleReferenceEvent : BaseIdentification
 {
+    [GraphQLIgnore]
     public uint EventIndex { get; init; }
     public string ModuleReference { get; init; } = null!;
     public AccountAddress Sender { get; init; } = null!;
@@ -36,6 +38,7 @@ public sealed class ModuleReferenceEvent : BaseIdentification
     public IList<ModuleReferenceRejectEvent> ModuleReferenceRejectEvents { get; init; } = null!;
     [GraphQLIgnore]
     public string? ModuleSource { get; private set; }
+    [GraphQLIgnore]
     public string? Schema { get; private set; }
     [GraphQLIgnore]
     public ModuleSchemaVersion? SchemaVersion { get; private set; }
@@ -116,6 +119,17 @@ public sealed class ModuleReferenceEvent : BaseIdentification
             return (versionedModuleSource, moduleSourceHex, moduleWasm);
         }
 
+        /// <summary>
+        /// The module can contain a schema in one of two different custom sections.
+        /// The supported sections depend on the module version.
+        /// The schema version can be either defined by the section name or embedded into the actual schema:
+        /// - Both v0 and v1 modules support the section 'concordium-schema' where the schema includes the version.
+        ///   - For v0 modules this is always a v0 schema.
+        ///   - For v1 modules this can be a v1, v2, or v3 schema.
+        ///- V0 modules additionally support section 'concordium-schema-v1' which always contain a v0 schema (not a typo).
+        /// - V1 modules additionally support section 'concordium-schema-v2' which always contain a v1 schema (not a typo).
+        /// The section 'concordium-schema' is the most common and is what the current tooling produces.
+        /// </summary>
         private static (string Schema, ModuleSchemaVersion SchemaVersion)? GetModuleSchema(WebAssembly.Module module, VersionedModuleSource moduleSource)
         {
             switch (moduleSource)
@@ -135,7 +149,7 @@ public sealed class ModuleReferenceEvent : BaseIdentification
                     {
                         return (moduleV1SchemaUndefined!, Application.Aggregates.Contract.Types.ModuleSchemaVersion.Undefined); // v1, v2, or v3
                     }
-                    if (GetSchemaFromWasmCustomSection(module, "concordium-schema-v1", out var moduleV1SchemaV1))
+                    if (GetSchemaFromWasmCustomSection(module, "concordium-schema-v2", out var moduleV1SchemaV1))
                     {
                         return (moduleV1SchemaV1!, Application.Aggregates.Contract.Types.ModuleSchemaVersion.V1); // v1 (not a typo)
                     }
@@ -149,7 +163,7 @@ public sealed class ModuleReferenceEvent : BaseIdentification
         {
             schema = null;
             var customSection = module.CustomSections
-                .SingleOrDefault(section => section.Name.StartsWith(entryKey, StringComparison.InvariantCulture));
+                .SingleOrDefault(section => section.Name.Equals(entryKey, StringComparison.InvariantCulture));
 
             if (customSection == null) return false;
             
@@ -184,6 +198,30 @@ public sealed class ModuleReferenceEvent : BaseIdentification
     [ExtendObjectType(typeof(ModuleReferenceEvent))]
     public sealed class ModuleReferenceEventExtensions
     {
+        private readonly ILogger _logger = Log.ForContext<ModuleReferenceEventExtensions>();
+
+        /// <summary>
+        /// Returns module schema in a human interpretable form. Only present if the schema is embedded into the
+        /// Wasm module.
+        /// </summary>
+        public string? GetDisplaySchema([Parent] ModuleReferenceEvent module)
+        {
+            if (module.Schema == null)
+            {
+                return null;
+            }
+
+            try
+            {
+                return InteropBinding.SchemaDisplay(module.Schema, module.SchemaVersion);
+            }
+            catch (Exception e)
+            {
+                _logger.Error(e, "Error when getting module schema to display");
+                return null;
+            }
+        }
+        
         [UseOffsetPaging(MaxPageSize = 100, IncludeTotalCount = true)]
         public IList<LinkedContract> GetLinkedContracts([Parent] ModuleReferenceEvent moduleReferenceEvent)
         {