+++ backend/Application/Aggregates/Contract/Extensions/ContractExtensions.cs
@@ -70,6 +70,7 @@ public static class ContractExtensions
 
         collection.AddSingleton<IContractJobRepository, ContractJobRepository>();
         collection.AddTransient<IContractJob, InitialModuleSourceCatchup>();
+        collection.AddTransient<IContractJob, UpdateModuleSourceCatchup>();
         collection.AddTransient<IContractJob, ParallelBatchBlockHeightJob<InitialContractAggregateCatchUpJob>>();
         collection.AddTransient<InitialContractAggregateCatchUpJob>();
     }