+++ backend/Tests/Aggregates/Contract/Entities/ModuleReferenceEventTests.cs
@@ -3,6 +3,7 @@ using System.IO;
 using System.Threading;
 using Application.Aggregates.Contract;
 using Application.Aggregates.Contract.Entities;
+using Application.Aggregates.Contract.Types;
 using Concordium.Sdk.Client;
 using Concordium.Sdk.Types;
 using FluentAssertions;
@@ -15,6 +16,34 @@ namespace Tests.Aggregates.Contract.Entities;
 public sealed class ModuleReferenceEventTests
 {
     [Theory]
+    [InlineData("cis1-wccd-embedded-schema-v0-unversioned.wasm", ModuleSchemaVersion.V0, false, 0)]
+    [InlineData("cis1-wccd-embedded-schema-v0-versioned.wasm.v0", ModuleSchemaVersion.Undefined, true, 0)]
+    [InlineData("cis2-wccd-embedded-schema-v1-unversioned.wasm.v1", ModuleSchemaVersion.V1, true, 1)]
+    [InlineData("cis2-wccd-embedded-schema-v1-versioned.wasm.v1", ModuleSchemaVersion.Undefined, true, 1)]
+    public async Task WhenCreateModuleSchema_ThenParseWithCorrectVersion(string fileName, ModuleSchemaVersion version, bool trim, int moduleVersion)
+    {
+        var client = new Mock<IContractNodeClient>();
+        var bytes = (await File.ReadAllBytesAsync($"./TestUtilities/TestData/{fileName}"));
+        if (trim)
+        {
+            bytes = bytes[8..];
+        }
+        VersionedModuleSource module = moduleVersion == 0 ? new ModuleV0(bytes) : new ModuleV1(bytes);
+        var queryResponseModuleSource = new QueryResponse<VersionedModuleSource>(BlockHash.From(new byte[32]), module);
+        
+        client.Setup(c => c.GetModuleSourceAsync(It.IsAny<IBlockHashInput>(), It.IsAny<ModuleReference>(), It.IsAny<CancellationToken>()))
+            .Returns(Task.FromResult(queryResponseModuleSource));
+        
+        // Act
+        var moduleSchema = await ModuleReferenceEvent.ModuleSourceInfo.Create(client.Object, 0, Convert.ToHexString(new byte[32]));
+
+        // Assert
+        moduleSchema.ModuleSource.Should().Be(Convert.ToHexString(bytes));
+        moduleSchema.ModuleSchemaVersion.Should().Be(version);
+    }
+    
+    
+    [Theory]
     [InlineData("module.schema_embedded.wasm.hex", "FFFF03010000000C00000054657374436F6E7472616374000000000001150200000003000000466F6F020300000042617202")]
     [InlineData("module.wasm.hex", null)]
     public async Task WhenCreateModuleSchema_ThenSchemaPresent(string fileName, string? schema)
@@ -26,7 +55,6 @@ public sealed class ModuleReferenceEventTests
         
         client.Setup(c => c.GetModuleSourceAsync(It.IsAny<IBlockHashInput>(), It.IsAny<ModuleReference>(), It.IsAny<CancellationToken>()))
             .Returns(Task.FromResult(queryResponseModuleSource));
-        var module = (await File.ReadAllTextAsync("./TestUtilities/TestData/module.wasm.hex")).Trim();
         
         // Act
         var moduleSchema = await ModuleReferenceEvent.ModuleSourceInfo.Create(client.Object, 0, Convert.ToHexString(new byte[32]));