+++ backend/CHANGELOG.md
@@ -1,7 +1,10 @@
 ## Unreleased changes
 
-## Unstable features
-- GraphQL endpoints for contracts and modules should currently be seen as unstable, indicating that the query models are likely to undergo changes.
+## 1.7.7
+- Added
+    - Endpoint to display module schema in a human interpretable form. The schema is present if it is embedded in the module source.
+- Bugfix
+    - Schema was incorrectly always mapped to undefined schema version. Fix implemented and job added which cleans up corrupted data.
 
 ## 1.7.4
 - Change contract- and module queries to use offset pagination.